pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ispeo"
version = "0.1.0"
description = "Inducible Squalene Production in E. coli ODE model"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "numpy>=1.23",
    "scipy>=1.9",
    "pandas>=1.5",
    "PyYAML>=6.0",
    "matplotlib>=3.6",
]

[project.scripts]
ispeo = "ispeo.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.2",
]
